// webview-ui/src/types.ts
export interface AutoloadEntry {
  namespace: string;
  path: string;
}

export interface AutoloadConfig {
  psr4: AutoloadEntry[];
  psr0: AutoloadEntry[];
  classmap: string[];
  files: string[];
}

export interface AutoloadData {
  autoload: AutoloadConfig;
  autoloadDev: AutoloadConfig;
}